enabling complex html as background slider
This appends a whole DOM element to the parallax mirror (utilizing the
same ".parallax-slider" class).

so far it only works properly when setting the naturalWidth and
naturalHeight options in the JS initialization. Somebody got a tip on
how to improve that?

I'm planning to implement a parallax carousel which should be possible
with that change.

also performance improvement suggestions from #71